transf: adjust the `while` lowering
Simply lowering `while cond` into:
```nim
while true:
  if not cond:
    break
  ...
```
does not work anymore, as `cond` would now (correctly) be part of the
`if`'s scope, not the `while`'s. To get around this without having to
use `if cond: body else: break`, which currently interferes with cursor
inference, `cond` is first assigned to a temporary.

Ideally, the `while` lowering would move back into `mirgen`, but
the existence of `closureiters` blocks that.