lowerings: re-use `result` field from environment
The `result` local in a coroutine needs to be redirected to the
`result` field of `Coroutine` instead of being mapped to a new field.

This is currently solved with a rather hack-y solution, where
`addField` and `getFieldFromObj` special case `skResult` symbols and
try to look them up by name (instead of id) first.

For searching the type, the existing `lookupInType` procedure is
moved from `semobjconstr` to `types`.