implement the coroutine transformation
The new `coroutines` module implements the coroutine transformations
(in a sense). It doesn't perform the main transformations itself,
instead acting as an orchestrator that delegates to `lambdalifting` and
`closureiters` for doing the heavy lifting.

The transformations is split into multiple steps:
1. turn the procedures marked with `.coroutine` into coroutine
   constructors, by turning their programmer-specified bodies into
   inner closure procedures
2. apply the `lambdalifting` pass on the constructor, which lifts the
   environment type (i.e., the coroutine instance type). The signature
   of the inner procedure is corrected, so that its header can already
   be passed to code generation
3. lastly, the `closureiters` pass is applied on the coroutine body and
   some post-processing takes place

A small amount of cooperation is required from `lambdalifting` for
coroutines, which are somewhere in-between normal procedure and closure
iterators (in terms of how they need to be processed).

Since much of the work is delegated, symbols and AST are temporarily
brought into semantically not valid states, but they are corrected
before leaving the pass again.