rework coroutine transformation
The previous implementation had one major flaw: it didn't account for
`closureiters` adding new fields to the environment type, which happens
when there's an `except` or `finally` in the body.

To properly support that situation, the pass is reworked to be bit less
clever:
* `coroutines` doesn't dispatch to `lambdalifting`
* instead, the inner procedure is a lambda expression
* the inner procedure is transformed during the core `transf` pass for
  the constructor (requires recursion, but is much easier)

When the `closureiters` is now run, the hidden parameter still has the
original type, allowing for the addition of new fields.