fix compiler crash when there's an error in the body
For coroutines with a result variables, the compiler crashes when
there's an error in the body.

This was because `addResult` didn't account for the result slot
existing but being empty, *appending* the result symbol to the AST,
thus leaving the actual result slot uninitialized and `buildErrorList`
then crashing due to a traversed node being nil.