semexprs: don't modify input AST on error
Eagerly updating the input AST in `builtinFieldAccess` prevents `x.y`
succeeding as `y(x)` when `x` is an error. The input AST is now only
modified (ideally it shouldn't be at all) when the LHS is not an
error, meaning that `y(x)` gets the untyped `x` when `x` turned out to
be an error.

In addition, `i` is renamed to `rhs`, making what it represents more
clear.