utils: update MIR pretty-printer
So that parameters and other locals can be displayed with their user-
defined names, an instance of the tree's corresponding `MirBody` has to
be passed along. A pointer is used so that it's still possible to not
pass along a body, in which case parameters and locals will be
displayed as raw IDs.