fix: VM crash with implicit object conversion
Summary
=======

Fix a bug with `cgirgen` where incorrect CGIR was produced for implicit
object conversions involving generic types, resulting in an internal VM
error when using the VM backend or running code at compile-time.

Details
=======

In `cgirgen.handleSpecialConv`, skipping was missing when testing for
whether the element type of a pointer-like type is an object type,
resulting in the conversion not being detected as an object conversion.
Code generators, such as `vmgen`, where the distinction between lvalue
and object conversion matters therefore produced wrong code.

`handleSpecialConv` is changed such that the types are skipped
properly. The object up- or down-conversion node now also uses the
correct line information.