vmjit: remove expression-related workaround (#1262)
## Summary

Make the MIR tree produced for standalone expressions (e.g., 
initializer of a `const`) syntactically well-formed. Previously, these
trees contained a trailing MIR expression, which required various
workarounds in other places (all of which are now removed).

## Details

* add the `exprToMir` procedure to `mirgen`; it produces a proper MIR
  body for standalone expressions by assigning the expression to the
  result variable
  * the body is wrapped in a scope (`mnkScope`), which means passes can
    now rely on the presence of an outermost scope block
* use `exprToMir` in `vmjit` for producing the MIR body for standalone
  expressions
* remove the dedicated code generation for standalone expressions from
  `vmgen` (`genExpr`)
* unify the expression/statement handling in `vmjit`
* remove the workarounds for trailing expressions from `cgirgen` and
  `mirpasses`

To replicate the previous behaviour, where no copy of the value was
introduced for lvalue expressions, `vmjit` wraps standalone lvalue
expressions in an `nkHiddenAddr`, prior to passing the AST to
`exprToMir`.