js: support variant destructors (#1267)
## Summary

Fix destructors not being properly executed for fields within case
objects when changing the active branch, and using the JS backend.

## Details

The destructor produced by `produceDestructorForDiscriminator` took a
reference to the *discriminant field* and used pointer arithmetic to
compute the address of the enclosing object. Not only is this
unnecessarily complex, it also doesn't work with the JS backend.

Instead, the destructor now takes a mutable reference to the
discriminant field's enclosing object type -- the MIR emitted for the
call is adjusted accordingly.

Since the destructor definitely modifies the object, `ekMutate` is used
instead of `ekInvalidate` for the parameter passing. This results in a
small move analyzer regression:
```nim
var x = Object()
var y = x.fieldOutsideOfVariant # not automatically moved anymore
x.kind = ... # potential branch switch that requires destruction
```

Finally, a test is added for ensuring proper destruction when changing
the active branch.