mir: split up `mnkConstr` (#1269)
## Summary

The `mnkConstr` node kind is removed and replaced with dedicated node
kinds for each construction (`array`, `seq`, `tuple`, and closure).
This is a step towards replacing type queries with syntax-only queries
as much as possible.

## Details

* `mnkConstr` is split into `mnkArrayConstr`, `mnkSeqConstr`,
  `mnkTupleConstr`, and `mnkClosureConstr`
* the proto-MIR covers `seq` construction now (`pirSeqConstr`)
* `mirgen` translates `pirSeqConstr` to `mnkSeqConstr`
* what `mnkConstr` usage is replaced with depends on the context 
* the pretty-printer uses a textual representation for the new node
  kinds that matches their high-level one (e.g.:
  `mnkArrayConstr` is rendered as `[...]`) 
* the `topt_no_cursor.nim` test is updated to account for the pretty-
  printer changes

It's now possible to search for, e.g., `seq` constructions
(`mnkSeqConstr`) without having to perform type queries on MIR nodes.