sigmatch: erroneous operands always trigger faux matches
Previously, only the absence of a type relation registered a match as
being a faux match. If an error type is bound to a generic parameter,
there an `isGeneric` relation, thus not registering a faux match.

This is fixed now. In addition, regressing from an `tyError` faux match
is prevented.