tests: fix `tcon1`
The third test case was failing, due to the results being swapped. The
underlying reason is that line was interpreted as
`test("from behind", test("and again", "more"))` due to the missing
closing parenthesis.

Since a faux match now counts as a use, the second overload had one
usage more than the first one, resulting in it being output first.
Adding the closing parenthesis fixes this.