tests: make `trecursive_table.nim` actually work
The `T` object type wasn't actually used anywhere in alive code,
meaning no code was generated for it, and thus that the regression it
intended to catch would have never been caught.

The `p` procedure is now exported, forcing code generation to run for
it (and subsequently the `T` type).