add the `mirtypes` module
In the future, it'll contain the type definitions for the type IR, but
for the moment, only the type environment type `TypeEnv` is and its
setup is located in it.

A instance of `TypeEnv` is directly embedded into `MirEnv`. Since
`TypeEnv` is a type requiring initialization, `MirEnv` too requires
explicit initialization, which is provided via `initMirEnv`.