prevent int/float literal types from being the canonical one
The first `PType` added to the environment for type X is treated as the
canonical one, which is a problem if said `PType` is an integer/float
literal type (all integer types would then show as integer/float
literal types).

To work around this, the `system.int` type is registered right away,
establishing it as the canonical `tyInt` type.