delay JIT creation in scripting context
Initializing the JIT environment requires a basic type environment, but
this is not present when using the VM evaluation pass, as the `system`
module hasn't been processed at the time the `PEvalContext` is
initialized.

To delay setting up the JIT environment as much as possible, the
evaluation pass is restructured:
* only a VM instance is assigned to the `ModuleGraph` at the start
* `setupGlobalCtx` creates a proper evaluation context from the
  instance
* evaluating statements, which now calls `setupGlobalCtx`, is prevented
  for declarative statements

This ensures that `initJit` is only called after the processing of the
`system` module has progressed far enough.