mir: use `TypeId` (#1277)
## Summary

Store references to types as `TypeId` instead of `PType`, making
`MirNode` a plain data-type. Besides speeding up the compiler, this
also paves the way to a data-oriented type IR for the MIR.

## Details

### Overview

* all MIR-level types use `TypeId` instead of `PType` (currently only
  `MirNode` and `Local`)
* the new `TypeEnv` type from the `mirtypes` module manages the type
  storage and mappings
* a `TypeEnv` instance is embedded in `MirEnv`
* registering a `PType` with `TypeEnv` yields a `TypeId`
* querying type information requires access to a `TypeEnv` instance

### Mapping Types

`TypeEnv` stores the `PType`-to-`TypeId` mappings in a `TypeTable`.
`TypeTable`, provided by the new `typemaps` module, is a custom
`Table` type that de-duplicates the `PType` keys by using structural
or symbolic comparison/hashing for the keys, thereby significantly
reducing the number of mappings and registered types.

At the moment, `PType` is still used as the type IR, so `TypeEnv` only
stores a `PType` for each `TypeId`.

The built-in types are mapped to statically known IDs, making some
translation and MIR construction easier.

### Constant expressions

To not require passing a `TypeEnv` to `datatables.getOrPut`, only the
type IDs are compared. This doesn't produce the same results as
`sameBackendType`, which ignores, for example, the distinct type
modifiers and aliases. As a result, `IntAlias(1)` and `int(1)` now
result in two `DataTable` entries instead of one.

Only using `TypeId` comparison/hashing does speed up `getOrPut`,
however.

### CGIR

* the CGIR still uses `PType`
* `cgirgen` turns the input `TypeId`s back into `PType`s
* since `cgirgen` creates new `PType` and assigns them to `Local`s,
  access to a *mutable* `MirEnv` is now required (in order to register
  the new `PType`s)

### VM backend

Packing constant data requires querying type information, and access to
a `TypeEnv` is thus required. A pointer to the `TypeEnv` instance is
passed along in `DataEncoder`.

### VM Evaluation Pass

Initializing the JIT environment requires a basic type environment,
which is not present when using the VM evaluation pass, as the `system`
module isn't yet processed at the time the `PEvalContext` is
initialized.

To delay setting up the JIT environment as much as necessary, the
evaluation pass is restructured:
* only a VM instance is assigned to the `ModuleGraph` at the start
* `setupGlobalCtx` creates a proper evaluation context from the
  instance
* evaluating statements, which now calls `setupGlobalCtx`, is prevented
  for declarative statements

This ensures that `initJit` is only called after the processing of the
`system` module has set up the basic type environment.

### Compiler Performance

`MirNode` being a plain data-type (no destructor) significantly speeds
up all MIR processing that creates new nodes or copies/moves them,
especially `TreeChangeset` application (~ -70%).

In addition, no copy/sink/destroy needing to be injected for `MirNode`s
also speeds up bootstrapping, since it means less work for the
destructor injection pass.

---------

Co-authored-by: Saem Ghani <saemghani+github@gmail.com>