transf: hoist bindings to beginning of `and`/`or`
`var`/`let` bindings within `and`/`or` expression are embedded into the
surrounding scope, which would cease to be the case after lowering, for
example, `x and y` into `(var tmp = x; (;if tmp: tmp = y); tmp), as
`y`'s lifetime is then delimited by the `if`'s scope.

So far, this problems was partially addressed by the "unscoped def"
handling in `cgirgen`, but only for locals not lifted into the
environment. Performing the hoisting in `transf` makes sure lifted
locals work as expected, and it also makes sure the code works
correctly without `cgirgen`.