mir: shrink size of `MirNode`
Summary
=======

Shrink the size of a `MirNode` from 24 byte (on a 64bit target) to 16
byte, reducing memory consumption and copying cost, as well as
increasing the number of nodes fitting in a cache line.

Details
=======

The `len` field was an int, resulting in the alignment requirement.
Since the `len` field is never negative, it's changed into a `uint32`
field. This also makes the size of a `MirNode` target-independent.