lower `new` and `ref` construction with MIR pass (#1280)
## Summary

Perform the lowering of `new` and `ref` constructions with a MIR pass,
moving the logic out of the C code generator. This is progress towards
having the code generator operate on only a very small core language.

## Details

### MIR

* introduce the `mnkRefConstr` node; `mnkObjConstr` is now only used
  for object constructions. While both are the same syntactically,
  they're very different semantically, and distinguishing them in
  syntax makes analysis and lowering easier
* no new CGIR counterpart is added; `mnkRefConstr` is translated to
  `cnkObjConstr`, meaning that no code generator change is required
* `TypeEnv` now provides the integer type to use for size-like values
  via the `sizeType` procedure

### Lowering

The lowering is largely equivalent to the one previously performed by
`cgen`. That is:
* the standard `new` is turned into `nimNewObjUninit` + `default(T)`
* the unsafe `new`-with-size is turned into `nimNewObj` + `default(T)`
* a `ref` construction (e.g.: `var x = RefObj(...)`) is turned into
  `def x = nimNewObjUninit(...); x[] = Obj(...)`

Compared to previous lowering in `cgen`, the difference is that both
the standard `new` and `ref` constructions now use the non-zeroing
`newNewObjUninit`. In the case of the standard `new`, if the type has
embedded type fields, this eliminates an unnecessary `memset`, but
observable behaviour stays the same.

### C Code Generator

Handling of the `mNew` magic (`genNew`) and `ref` constructions
(implemented in `genObjConstr`) are both obsolete, and thus removed.