cgirgen: restore the `inUnscoped` workaround
It's still needed for temporaries requiring destruction that are
spawned within the right-hand operand of `and`/`or` expressions to
work.

`mirgen` could (and likely should) simply wrap the right-hand
operand in its own scope. This does change observable behaviour,
however, and is thus better done as a follow up PR (if at all).