cgen: don't emit excessive strings constants (#1285)
## Summary

For the C backend, don't emit excessive, duplicate string constants
for field checks when compiling with `--declaredlocs:off` (the
default).

## Details

The code generator handling for field checks unconditionally emitted
raw strings, resulting in a separate string constant being emitted for
each field check appearing in the code. This is unnecessary when the
error message string value stays as is, bloating the produced C code.

If the error message string does not need to be modified (i.e., when
`--declaredlocs` is disabled), the cache-using overload of
`genStringLiteral` is now used.