mirpasses: implement string COW handling with MIR pass (#1287)
## Summary

Perform the injection of runtime procedure calls for handling copy-on-
write strings with an MIR pass, instead of doing it in the C code
generator, removing another responsibility from the latter.

## Details

The new MIR pass does the equivalent of what `cgen` previously did,
that is, injecting `nimPrepareStrMutationV2` calls wherever a string's
underlying storage might be modified.

The injection logic, together with the `lfPrepareForMutation` flag, is
removed from `cgen`.