cgen: better code generation for `.noreturn` calls
Summary
=======

Improve C code generation for `.noreturn` calls, by omitting the
unnecessary error flag check for them.

Details
=======

If the procedure doesn't return via non-exceptional control flow, the
`if(NIM_UNLIKELY(*nimErr_))` guard guarding the unwinding is
unnecessary, since it's guaranteed that the error flag is set (all
exits of the procedure are due to exceptional unwinding).

In addition, the `canRaiseConservative` usage is unnecessary too, since
a `cnkCheckedCall` already signals that the call can raise. The
procedure has no more users and is thus removed.

Finally, for compilers that support it, a call to a `.noreturn`
procedure that doesn't raise is now followed by an `__assume(0)`
statement, communicating to the C compiler that control-flow never
falls out of the procedure.