mirgen: better code generation for `high` (#1286)
## Summary

Omit unnecessary to-slice conversions for `high(x)` calls where `x` is
a `seq` value, resulting in a small compiler speed up. All backends are
affected.

## Details

There's no dedicated `high` overload for `seq`, so the `openArray`
overload is picked for `seq` arguments. The implicit argument
conversion was previously translated as is, resulting in an extra
temporary and conversion operation, both which are unnecessary, since
all three code generators support `high` for `tySequence` operands.

While an optimizing C compiler is able to eliminate the conversion,
doing it on the NimSkull side still benefits debug builds and the
JavaScript and VM target, as well as reduce the workload for the MIR
passes and code generators.

### Tests

* add a new test category for optimization-related tests
* add a test to ensure there's no conversion for `high(seq)` calls