cgen: better code generation for `.noreturn` calls (#1288)
## Summary

Improve C code generation for `.noreturn` calls, by omitting the
unnecessary error flag check for them.

## Details

If a procedure only returns due to exceptional unwinding, the
`if(NIM_UNLIKELY(*nimErr_))` guard guarding the unwinding is
unnecessary, since it's guaranteed that the error flag is set (all
exits of the procedure are due to exceptional unwinding).

Unconditionally jumping to the error handling target makes it clear
to the C compiler that control-flow doesn't continue normally after
the call

### Other improvements

In addition, the `canRaiseConservative` usage is unnecessary too, since
a `cnkCheckedCall` already signals that the call can raise. The
procedure has no more users and is thus removed.

Finally, for compilers that support it, a call to a `.noreturn`
procedure that doesn't raise is now followed by an `__assume(0)`
statement, communicating to the C compiler that control-flow never
returns to the caller.