system: mark `raiseX` runtime procedures as noreturn (#1289)
## Summary

Mark the `raiseX` runtime procedures used by the C backend as
`.noreturn`, to allow for generation of more efficient code.

## Details

For less code duplication, the `.noinline`, `.noreturn`, and
`.compilerproc` pragmas are combined into the custom `.errorPrc`
pragma.

At the moment, code generation is not affected, since `cgen` itself
emits calls to the `raiseX` procedures, but this is going to change
once MIR passes perform lowering of the runtime checks.