mirpasses: lower `mNew` and `mWasMoved` with MIR pass
Instead of in the code generator(s), lower the magics with a combined
MIR pass. This also fixes `move` calls not resetting the source
location when using the VM backend (`vmgen` implemented the magic
improperly). `jsgen` had a similar bug, where locations of primitive
types were not reset properly -- this is fixed too.