refactor(pragmas): improve `processCompile` (#1294)
## Summary

Refactor the `pragmas.processCompile` procedure to not mutate input
AST, and add a specification test for the undocumented pattern-matching
version of `.compile`.

## Details

For `pragmas.processCompile`:
* don't mutate input AST
* use proper error propagation (i.e., keep the shape of the AST intact)
* test that the specified path is relative before trying to interpret
  it as a relative path (less work)

The `.compile` pragma has a pattern-matching version, but this was
not documented anywhere. A specification test derived from the current
behaviour is added for it.