prevent collisions with object files for NimSkull modules
It's possible to add to the build external C files that use names for
which the generated object file would collide with that of a NimSkull
module's object file. To prevent this from happening, external C files
are now compiled to object files located in the `external` sub-
directory within the cache directory.