fix(vmgen): string literal changing at run-time
Summary
=======

Fix a string literal's underlying storage being modified when passing
a string literal to a `sink string` parameter, where the parameter is
modified within the callee. Only the VM backend was affected.

Details
=======

No copy of the string literal (which was lifted into a constant
earlier) is created for string literals passed to `sink` parameters.
This is not a problem for the C (copy-on-write strings are used) and
JS backend (a new run-time string instance is created on each string
literal usage), but it is for the VM backend.

Without a copy, the callee can directly modify the constant string's
underlying character storage, affecting all string literal usage
throughout the program.

`mirgen` should introduce an intermediate temporary for the argument in
this case, but - at the moment - doing so would lead to worse code
generation for all code generators, so the problem is worked around in
`vmgen`, by creating a copy of the string constant when passed to a
`sink string` parameter.