semtempl: always bind gensyms in field positions
This lifts the restriction that a gensym'ed symbol cannot be used in a
field access position. However, at the moment, it also causes symbols
such as `skVar` symbols to be bound there, which is incorrect.

Since gensyms don't reach into `semTemplSymbol` anymore, all
conditionals including them become false and are thus obsolete.