semtempl: only consider a subset of symbols in field positions
Only routine symbols and type / generic parameter symbols are
considered in field access positions, addressing the problem that
symbols of, e.g., locals blocked other symbols.

Symbol binding in generics for dot expressions works similar, with
the differences that:
* types and generic parameters are not considered there
* a *closed* symbol choice is forced, instead of an *open* one

Using an open symbol choice is wrong (it violates the specification),
but is required for backwards compatibility.