fix `tcustom_numeric_literals` test
The field part of a dot expression may now be a symbol choice, which
interferes with rendering of custom numeric literals (e.g., `1'lit`),
where the special rendering was only enabled for `nkIdent` and `nkSym`
nodes.

To keep the test working, `getPIdent` (which is used by `isCustomLit`)
now also returns the identifier for symbol-choice nodes.