fix the `tcustom_numeric_lit` test properly
Just retrieving the identifier from the symbol choice doesn't work,
since then the symbol-choice node is passed to `gident`, where `atom`
then crashes because symbol-choices are not allowed (a symbol-choice
is, not necessarily, an atom, so this is correct).

Instead of `getPIdent` retrieving an identifier from a symbol-choice,
`isCustomLit` now special-cases symbol-choices.