fix(testament): batching being ignored for megatest (#1299)
## Summary

Fix batching being ignored for the megatest. Only tests part of the
current batch are now joined into the megatest.

Fixes https://github.com/nim-works/nimskull/issues/1037

## Details

In order to not duplicate the computation of whether the test is in the
current batch, a pseudo `TTest` instance is created that's then passed
to `computeEarly`.