threads: default `createThread` returns a value
`createThread` returning a value instead of using in-place construction
is the better interface, preventing leaks when calling `createThread`
on the same location multiple times. This interface was previously not
possible, since the `Thread` instance had to be at fixed memory
location.

For backwards compatibility, an in-place version of `createThread` is
added.