threads: improve interface and fix bugs (#1302)
## Summary

Make `Thread` a proper thread handle, fixing multiple bugs and making
the interface more robust:
* a thread is now officially allowed to outlive the `Thread` instance
  (which would previously lead to crashes)
* a `Thread` handle for a running thread can now be safely moved around
* there's now a `createThread` overload that returns a `Thread`
  instance
* `Thread.finished` having a race condition

## Details

The management-data for a thread was previously part of the `Thread`
type itself, with a pointer to the `Thread` instance passed to the
internal thread procedure, requiring the `Thread` instance for a
running thread to always stay at the same memory location.

As a consequence of this design, moving the `Thread` instance, or the
`Thread` instance going out of scope before the thread is finished,
resulted in undefined behaviour.

To address this problem, the management-data is made part of a heap-
allocated structure (`ThreadCore`), which is then shared between the
handle and spawned thread. The type uses manual reference counting in
order to support detached threads (threads that outlive their handle).

Since a `Thread` instance can now be moved around in memory, the base
`createThread` procedure can now return a `Thread`, which is the better
interface, as it prevents leaks caused by forgetting to destroy the
`Thread` instance first. For backwards compatibility, an in-place
`createThread` overload is added.

### Tests

Two new tests are added:
* a test for ensuring that the thread's internal data is freed
  correctly
* a test for ensuring that the thread parameter is destroyed

The pre-existing `tthread_destroy_before_finish.nim` test had a few
issues, which are now fixed:
* it had no proper description
* it wasn't guaranteed to actually fail (in the edge-case that the
  spawned thread finishes before the `reset` call)
* access to the `ok` global was not atomic (resulting in a race
  condition)
* it unnecessarily imported `std/os`