add the `supportsZeroMem` type trait
Summary
=======

The type is intended for libraries to figure out whether `zeroMem`
leaves a location in it's "default-initialized" state, something that's
useful to know for optimization or correctness purposes in generic
code.

Details
=======

Without the type trait, either an `is` approximation or type-inspection
macro would have to be used.

The only types not supporting `zeroMem` for initialization are:
* types that don't have a zero-default (such as `range` types, enums
  not starting at zero, `.requiresInit` types, not-nil pointers, etc.)
* object types that have a type header field (non-pure object types
  that can be inherited from)
* aggregate types containing the former two

The `nimskullHasSupportsZeroMem` conditional symbol is added so that
support for type-trait can be detected at compile-time.